Delay effect updated. Ran stable for a ~15 min test. Note: delay buffer might get filled with noise from the guitar (not from playing), that lasts for the length of the delay buffer. Not a bug, but an issue that should be adressed.

+++ paEffex_Win/DelayEffect.h
@@ -7,21 +7,21 @@ public:
 	DelayEffect();
 	~DelayEffect();
 //	int applyEffect(double * iData, double * oData, int bufferSize);
-	int applyEffect(int * iData, int* oData, int bufferSize);
+	int applyEffect(float * iData, float* oData, unsigned long bufferSize);
 	int changeEffectParameters(void * userdata);
 
 	void set_gain_(double gain);
 	void set_dlyTime_(double dlyTime);
-	int get_gain_();
-	int get_dlyTime_();
+	unsigned long get_gain_();
+	unsigned long get_dlyTime_();
 
 private:	
-	int dlyBufferSize_;
-	int * dlyBuffer_; // Buffer with samples for delay processing
-	int sample_cntr_;
-	double dlyGain_; // Delay gain
-	double dlyTime_; // MS between each repeat
-	int dlySamples_; // delay time in samples
-	int sampleRate_;
+	unsigned long dlyBufferSize_;
+	float * dlyBuffer_; // Buffer with samples for delay processing
+	unsigned long sample_cntr_;
+	float dlyGain_; // Delay gain
+	float dlyTime_; // MS between each repeat
+	unsigned long dlySamples_; // delay time in samples
+	unsigned long sampleRate_;
 };
