+++ paEffex_Win/IGuitarEffect.h
@@ -5,7 +5,7 @@ class IGuitarEffect
 public:
 	IGuitarEffect();
 //	virtual int applyEffect(double* iData, double* oData, int bufferSize) = 0;
-	virtual int applyEffect(int* iData, int* oData, int bufferSize) = 0;
+	virtual int applyEffect(float* iData, float* oData, unsigned long bufferSize) = 0;
 	virtual int changeEffectParameters(void * userdata) = 0;
 	virtual ~IGuitarEffect();
 	bool enableEffect_;
